requirements.txt
django==3.1.3
djangorestframework==3.12.2
requests==2.25.0
